Main

Description

Introdução

A tarefa 5 constitui a tarefa central do projeto, pois é esta que liga todas as outras e nos permite visualizar o jogo. Esta tarefa utiliza a biblioteca “gloss” que permite representar graficamente o mapa, os jogadores e o decorrer do jogo. A principal dificuldade desta parte do projeto foi ter de aprender a usar o “gloss”, que é diferente daquilo que estávamos habituados em haskell.

Objetivos

Todos os elementos importantes para o desenho do jogo num determinado instante estão incluídos no EstadoGloss. O primeiro objetivo que nós definimos foi desenhar um mapa usando a função “gera” da tarefa 1. A nossa abordagem passou por definir um “x” e um “y” iniciais que corresponderiam ao canto superior esquerdo do mapa (ponto (x,y)). Desse ponto (x,y), a primeira linha corresponderia a traduzir a primeira peça para a imagem correspondente e somar ao “x” o comprimento da imagem. Em seguida desenha-se a próxima peça aplicando uma translação para a direita correspondente ao comprimento da imagem da peça anterior. Este processo repete-se até acabar de desenhar a linha. A próxima linha seria desenhada aplicando uma translação para baixo correspondente à altura da imagem de uma peça. As rampas são desenhadas como polígonos pois este método é mais eficaz e produz melhores resultados que adicionar imagens individuais para cada tipo de rampa. Deste modo já era possível desenhar qualquer mapa. A “seed” do mapa é gerada aleatoriamente ao abrir o jogo.

Em seguida desenhamos os jogadores, tarefa que foi facilitada pela existência dos tais “x”, “y” iniciais. Ao “x” inicial adicionamos a distancia percorrida pelo jogador e o jogador já é desenhado na posição correta. Para a altura, fazemos uma translação para cima correspondente à altura do jogador. Aqui foi preciso criar uma função que devolve a altura das peças para podermos fazer a referida translação quando o jogador está no estado “Chao”. A função altura peça também serve para adicionar altura às peças do mapa, fazendo também uma translação vertical de cada peça pela sua altura. Depois de termos o mapa e os jogadores a funcionar tivemos de adicionar à função “reageTempo” a tarefa 4, que corresponde à passagem do tempo. A tarefa 2 foi adicionada na função reageEvento, o que permitiu controlar os jogadores e efetivamente jogar.

Um problema surgiu neste ponto, ao abrir o jogo num computador com maior resolução o mapa ficava cortado. Assim foi preciso implementar um mecanismo de “scalling”. Foi definida uma função “comp :: Int -> Float ” que dado o comprimento e a altura de uma janela dava o comprimento que as peças individuas deveriam ter. Implementado este mecanismo a janela do jogo já podia ser redimensionada e o jogo responderia a este redimensionamento. O passo seguinte foi desenvolver um menu de jogo que inclui três opções : um jogador, dois jogadores e jogar contra o bot da tarefa 6. Foi adicionado ao “EstadoGloss” um inteiro que correspondia ao botão clicado. O “bot” foi adicionado na função “reageTempo”, onde a cada instante de tempo o “bot” faz uma decisão. Foi adicionado um cronómetro no canto superior direito que mostra o tempo decorrido.

Por último, foram adicionadas as mensagens de vitória que indicam o jogador vencedor e o tempo realizado por este e foi modificada a função “desenhaMapa” para passar a desenhar os jogadores intercalados com o mapa de modo que o jogador não fique sempre acima do mapa, mesmo quando existe algo à sua frente.

Conclusão

Esta tarefa colocou-nos diante de vários desafios que nos obrigaram a sermos criativos para os solucionar. A nossa tarefa apresenta-se como um jogo funcional onde é possível jogar sozinho, jogar com outra pessoa e jogar contra um “bot”. O jogo reage ao tamanho da janela, ajustando-se convenientemente e os mapas são aleatoriamente gerados no inicio do jogo. Podemos considerar esta tarefa como a mais importante do jogo, pois é nesta tarefa onde acontece a magia de transformar código em algo com que podemos interagir.

Documentation

type EstadoGloss = (Float, Float, Estado, [Picture], Menu, Int, (Int, Int), Int)

Estado que contem as caracteristicas necessárias á renderização do jogo

data Menu

Informação sobre os menus

Constructors

Inicial

Pausa

NoMenu

xWindow :: Int

Valor predefinido da largura da janela inicial

yWindow :: Int

Valor predefinido da altura da janela inicial

comp :: Int -> Float

Comprimento individual de cada peça dada a largura da janela

criaEstado :: Int -> Int -> Int -> Int -> Estado

Função que cria o estado inicial do jogo atravês dos parametros estabelecidos pelo utilizador

criaJog :: Mapa -> Int -> Int -> [Jogador]

Função auxiliar da função "criaEstado" que cria os a lista de jogadores a ser implementada no estado inicial do jogo

estadoInicial :: [Picture] -> Int -> EstadoGloss

Função que cria o EstadoGloss inicial do jogo

desenhaEstado :: EstadoGloss -> Picture

Função que apresenta no ecrã o estado atual do jogo

desenhaListaJogador :: (Int, Int) -> Int -> Mapa -> [Jogador] -> [Picture] -> [Picture]

Função auxiliar da função "desenhaEstado" que possiblita a vizualização de determinada lista de jogadores

desenhaJogador :: (Int, Int) -> Int -> Mapa -> Jogador -> [Picture] -> Picture

Função auxiliar da função "desenhaListaJogadores" que desenha individualmente um jogador

desenhaMapa :: (Int, Int) -> Mapa -> Mapa -> Float -> Float -> [Jogador] -> [Picture] -> [Picture]

Função auxiliar da função "desenhaEstado" que possiblita a vizualização de determinado mapa

desenhaLinha :: Int -> [Peca] -> Float -> Float -> [Picture] -> [Picture]

Função auxiliar da função desenhaMapa que desenha as pistas do mapa

desenhaPeca :: Int -> Peca -> Float -> Float -> [Picture] -> Picture

Função auxiliar da função desenhaMapa que desenha as peças do mapa e os poligonos abaixo destas

auxDesenhaPeca :: Int -> Peca -> Color -> Color -> Float -> Float -> Picture

Função que desenha rampas e os poligonos abaixo destas

reageEvento :: Event -> EstadoGloss -> EstadoGloss

Função que calcula a jogada efetuada atravês dos diferentes eventos do teclado

selecionaJogada :: Event -> Estado -> Int -> Estado

Funçãu auxiliar da funcção reageEvento que seleciona a jogada a efetuar

reageTempo :: Float -> EstadoGloss -> EstadoGloss

Função que altera o estadoGloss do jogo mediante a passagem do tempo

vencedor :: [Jogador] -> Mapa -> Int -> Int

Função que devolve o jogador vencedor. 0 se nenhum venceu ainda , 1 se foi o primeiro e 2 se foi o segundo e 3 se a lista de jogadores for vazia

auxTempo :: Maybe Jogada -> Estado -> Estado

Função auxiliar que converte as jogadas do bot num estado

alteraEstado :: Float -> Estado -> Estado

Função auxiliar da função reageTempo que altera o estado de acordo com a passagem do tempo

alteralj :: Float -> Estado -> [Jogador]

Função auxiliar da função altera estado que é aplicada apenas às listas de jogadores

fr :: Int

Frame rate

dm :: Display

Display

displaySize :: (Int, Int)

Tamanho do display

main :: IO ()

Função principal da Tarefa 5.

NB: Esta Tarefa é completamente livre. Deve utilizar a biblioteca gloss para animar o jogo, e reutilizar de forma completa as funções das tarefas anteriores.