Tarefa2_2019li1g004

Description

Este módulo define funções comuns da Tarefa 2 do trabalho prático.

Testes

testesT2 :: [(Int, Jogada, Estado)]

Testes unitários da Tarefa 2.

Cada teste é um triplo (identificador do Jogador,Jogada a efetuar,Estado anterior).

Funções principais da Tarefa 2.

jogada

Arguments

:: Int
O identificador do Jogador que efetua a jogada.

-> Jogada
A Jogada a efetuar.

-> Estado
O Estado anterior.

-> Estado
O Estado resultante após o jogador efetuar a jogada.

Efetua uma jogada.

alteraJogador :: Jogador -> Jogada -> Mapa -> Jogador

Função que altera um jogador consoante a jogada selecionada

alteraPistaJogador :: Jogador -> Jogada -> Mapa -> Jogador

Função auxiliar da função alteraJogador

Altera o jogador a nível da pista em que se encontras

testaMudanca :: Jogador -> Jogada -> Mapa -> Int

Função auxiliar de alteraPistaJogador

Testa o que acontece quando uma mudança de pista é efetuada

hJogador :: Jogador -> Mapa -> Double

Função que calcula a altura atual do jogador

hAdjacente :: Jogador -> Jogada -> Mapa -> Double

Função que calcula a altura do jogador na peça de mudança pretendida

hPeca :: Int -> Double -> Mapa -> Double

Função que calcula a altura total da peça onde se encontra o jogador atravês da distancia percorrida e pista onde se encontra

alteraInclinacao :: Jogador -> Jogada -> Jogador

Função que altera a Inclinação do Jogador

colocaCola :: Jogador -> Mapa -> Mapa

Função que altera o mapa quando a jogada selecionada é Dispara

alteraTipoPecaCola :: Peca -> Peca

Função auxiliar da coloca cola que altera a peça pretendida ,no caso em que, se encontra imediatamnete antes do jogador que efetua o disparo