Tarefa3_2019li1g004

Description

Introdução

O objetiva desta Tarefa é comprimir ao máximo um Mapa de forma a que o resultado dessa compressão seja um conjunto de instruções que sendo executadas por bulldozers contruam o mapa previamente comprimido.

Objetivos

Inicialmente o objetivo foi traduzir o mapa de forma a que cada peça do mesmo correspondesse a uma instrução, posteriormente esse conjunto de instruções teria de ser comprimido de forma a reduzir a lista de instruções a ser efetuada. Depois da tradução comprimimos verticalmente o mapa comparando os tipos das instruções referentes ao fim de uma pista e ao ínicio de outra, quando estas são iguais ocorre compressão em instruções aplicadas às duas pistas. Por fim, realiza-se a compressão horizontal que usa o tipo repete para diminuir o número de instruções. Assim quando n intruções aplicadas à mesma pista são semelhantes estas são comprimidas em apenas uma instrução do tipo repete n, que repete n vezes a mesma instrução. Para isso comparamos pista a pista as instruções semelhantes que ocorriam de seguida, agrupando-as posteriormente.

Discussão e Conclusão

Como resultado, obtivemos uma taxa de compressão não muito boa a nível vertical, podendo ser melhorada caso comparassemos não só o fim e o início de duas pistas adjacentes mas sim todas as pistas nas várias posições relativas, mas tivemos problemas ao faze-lo a nível da ordem de leitura das instruções, o que provocava erros na leitura do mapa. A nível horizontal a compressão demonstrou ser bastante satisfatória e fiavél já que comprimiu o possível segundo o oráculo e devolveu o mapa correto em todos os testes efetuados.

Testes

testesT3 :: [Mapa]

Testes unitários da Tarefa 3.

Cada teste é um Mapa.

Funções principais da Tarefa 3.

desconstroi :: Mapa -> Instrucoes

Desconstrói um Mapa numa sequência de Instrucoes.

NB: Uma solução correcta deve retornar uma sequência de Instrucoes tal que, para qualquer mapa válido m, executar as instruções '(desconstroi m)' produza o mesmo mapa m.

NB: Uma boa solução deve representar o Mapa dado no mínimo número de Instrucoes, de acordo com a função tamanhoInstrucoes.

traducaoMapa :: Mapa -> Int -> Instrucoes

Dado um mapa converte-o em instruções

traducaoPista :: [Peca] -> Int -> Instrucoes

Dado uma pista converte-a em instruções

traducaoPeca :: Peca -> Int -> Instrucao

Dada uma peça converte-a numa instrução

agrupaV :: Instrucoes -> Instrucoes

Agrupa verticalmente padrões semelhantes

agrupah :: Instrucoes -> Instrucoes

Agrupa horizontamente instrucoes semelhantes

somaI :: Instrucao -> Instrucao -> Instrucao

Soma as listas de duas Instruções do mesmo tipo

somaListaI :: Instrucoes -> Instrucoes

Soma uma lista de instruções

insTest :: Instrucao -> Instrucao -> Bool

Testa se duas instruções são do mesmo tipo

nPistas :: Instrucoes -> Int

Dada uma lista de Instruções dá o número de pistas

subList :: Eq a => [a] -> [a] -> Bool

Testa se uma lista está contida noutra

getPista :: Instrucao -> Int

Função que dada uma Instrução dá a primeira pista em que a mesma se aplica

separaPorPista :: Int -> Instrucoes -> Instrucoes

Função que dada uma lista de Instruções e uma determinada pista separa todas as instruções que se aplicam a essa pista