Tarefa4_2019li1g004

Description

Este módulo define funções comuns da Tarefa 4 do trabalho prático.

Testes

testesT4 :: [(Double, Mapa, Jogador)]

Testes unitários da Tarefa 4.

Cada teste é um par (tempo,Mapa,Jogador).

Funções principais da Tarefa 4.

passo

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após um passo.

Avança o estado de um Jogador um passo em frente, durante um determinado período de tempo.

acelera

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após acelerar.

Altera a velocidade de um Jogador, durante um determinado período de tempo.

move

Arguments

:: Double
O tempo decorrido.

-> Mapa
O mapa utilizado.

-> Jogador
O estado anterior do Jogador.

-> Jogador
O estado do Jogador após se movimentar.

Altera a posição de Jogador, durante um determinado período de tempo.

testaTransicao :: Jogador -> Mapa -> Double -> Jogador

Função que avalia a peça para onde o jogador transitou e a peça anterior

alturaPeca :: Int -> Double -> Mapa -> Double

inclinacaoPeca :: Int -> Double -> Mapa -> Double

Função que calcula a inclinação da Peca para onde o jogador transita

atritoPeca

Arguments

:: Int
A pista do jogador

-> Double
A posição no eixo Ox do jogador

-> Mapa
O mapa atual

-> Double
O atrito

Devolve o valor do atrito da peça

pisoPeca :: Peca -> Piso

Dada uma peça devolve o piso correspondente

returnX :: Ponto -> Double

Função que dado um ponto cartesiano devolve a abcissa desse ponto

returnY :: Ponto -> Double

Função que dado um ponto cartesiano devolve a ordenada desse ponto