Tarefa6_2019li1g004

Description

Introdução

Esta tarefa consistiu na criação de um bot, isto é, na criação de uma inteligência artificial primitiva capaz de jogar o jogo automaticamente e de forma possivelmente mais eficaz que um humano.

Objetivos

Como principal objetivo desta Tarefa assumimos o tempo, ou seja, tentamos minimizar o tempo necessário para concluir um mapa ou percorrer determinado número de peças, criando assim um bot semi passivo que prioriza o seu avanço em relação à posição dos oponentes. Com isto queremos dizer que o robot apenas dispara quando se encontra na posição ideal, posição esta que é verificada atravês da função passo da tarefa 4. A cada jogada o robot calcula a distância que percorrerá até à próxima jogada em diferentes pistas. Caso a melhor posição para o jogador seja numa pista diferente da que se encontra no momento da jogada, o bot seleciona a jogada necessária para que o jogador se mova para a melhor posição (considerando que a mudança pode ser mortal), caso contrário o robot opta por um comportamento mais agressivo e Dispara. Quando o jogador se encontra no ar o robot calcula qual será a peça de embate com o solo e consecutivamente a sua inclinação , determinado assim as jogadas necessárias para não morrer no embate devido à sua inclinação.

Discussão e conclusão

A nível de resultados, obtivemos um bot que em não morreu em nenhum dos testes localmente efectuados e que torna quase impossivel a vitória de um concorrente humano. Demonstra uma grande capacidade de evitar obstáculos que abrandem a sua progressão e usufrui sempre dos blocos boost espalhados pelas diferentes pistas.

Funções principais da Tarefa 6.

bot

Arguments

:: Int
O identificador do Jogador associado ao robot.

-> Estado
O Estado para o qual o robot deve tomar uma decisão.

-> Maybe Jogada
Uma possível Jogada a efetuar pelo robot.

Define um robot capaz de jogar autonomamente o jogo.

retJog :: Int -> Estado -> Jogador

Função que dado um estado e o identificador de Jogador devolve o jogador desejado

distJog :: Jogador -> Double

Função que dado um jogador devolve a distancia percorrida pelo mesmo

bestOf3 :: (Jogador, Int) -> (Jogador, Int) -> (Jogador, Int) -> Int

Função que sendo fornecidos 3 posições de um Jogador devolve o que se encontra em melhor posição

retardMorto :: Jogador -> Jogador

Função que retarda a posição do jogador para que o estado de tempo em que se encontra morto seja penalizado pela função verfPista

verfPista :: Jogador -> Mapa -> Maybe Jogada

Função que verifica qual a melhor posição do Jogador e devolve a jogada a ser efetuada

verfInc :: Jogador -> Mapa -> Maybe Jogada

Função que verifica qual a melhor jogada quando o jogador se encontra no ar

aterragemX :: Jogador -> Mapa -> Double

Função que calcula peça de aterragem do jogdador que se encontra no ar